Sales/requirements.txt
sentence-transformers==2.5.1
scikit-learn==1.4.1.post1
xgboost==2.0.3
matplotlib==3.8.3
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.1
torch==2.2.1
transformers==4.38.2
openai==1.12.0
python-multipart==0.0.9
shap==0.44.1
plotly==5.19.0 